Comments 4-6
Addressed comments 4-6. Utilized the discussion of written prompts as qualitative, continuous measures (comment 5) to motivate the usage of LSA (comment 6). Provided a fuck-ton of references for the usage of LSA in the literature (comment 6, see page 8). Streamlined the end of introduction, clarified hypothesis, and motivated our methodology. Added clarification sentence regarding the "hobbies-and-interest" prompts. MAY STILL NEED MORE CONVERSATION REGARDING VALIDITY OF WRITTEN PROMPTS AS A MEASURE OF ATTRACTIVENESS!